Fix IDE low version exception
