+++ src/main/java/com/shuzijun/leetcode/plugin/actions/ResetTimeAction.java
@@ -1,17 +1,16 @@
 package com.shuzijun.leetcode.plugin.actions;
 
 import com.intellij.openapi.actionSystem.AnActionEvent;
-import com.intellij.openapi.wm.WindowManager;
 import com.shuzijun.leetcode.plugin.model.Config;
+import com.shuzijun.leetcode.plugin.model.Question;
 import com.shuzijun.leetcode.plugin.timer.TimerBarWidget;
 
 /**
  * @author shuzijun
  */
-public class ResetTimeAction extends AbstractAsynAction {
+public class ResetTimeAction extends AbstractTimeAction {
     @Override
-    public void perform(AnActionEvent anActionEvent, Config config) {
-        TimerBarWidget timerBarWidget = (TimerBarWidget) WindowManager.getInstance().getStatusBar(anActionEvent.getProject()).getWidget(TimerBarWidget.ID);
+    public void perform(AnActionEvent anActionEvent, Config config, TimerBarWidget timerBarWidget, Question question) {
         timerBarWidget.reset();
     }
 }
