+++ src/main/java/com/shuzijun/leetcode/plugin/actions/StopTimeAction.java
@@ -1,17 +1,16 @@
 package com.shuzijun.leetcode.plugin.actions;
 
 import com.intellij.openapi.actionSystem.AnActionEvent;
-import com.intellij.openapi.wm.WindowManager;
 import com.shuzijun.leetcode.plugin.model.Config;
+import com.shuzijun.leetcode.plugin.model.Question;
 import com.shuzijun.leetcode.plugin.timer.TimerBarWidget;
 
 /**
  * @author shuzijun
  */
-public class StopTimeAction extends AbstractAsynAction {
+public class StopTimeAction extends AbstractTimeAction {
     @Override
-    public void perform(AnActionEvent anActionEvent, Config config) {
-        TimerBarWidget timerBarWidget = (TimerBarWidget) WindowManager.getInstance().getStatusBar(anActionEvent.getProject()).getWidget(TimerBarWidget.ID);
-        timerBarWidget.stopTimer();
+    public void perform(AnActionEvent anActionEvent, Config config, TimerBarWidget timerBarWidget, Question question) {
+        timerBarWidget.reset();
     }
 }
